+++ src/Auth/GuardDirective.php
@@ -59,10 +59,9 @@ GRAPHQL;
 
         $fieldValue->setResolver(function ($root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($previousResolver) {
             // TODO remove cast in v6
-            $with = (array) (
-                $this->directiveArgValue('with')
-                ?? [AuthServiceProvider::guard()]
-            );
+            $with = (array)
+                $this->directiveArgValue('with', AuthServiceProvider::guard())
+            ;
 
             $this->authenticate($with);